+++ modules/cpu/lights/SpotLight.cpp
@@ -14,15 +14,15 @@ void *SpotLight::createIE(const void *instance) const
   ispc::SpotLight_set(ie,
       (const ispc::vec3f &)position,
       (const ispc::vec3f &)direction,
-      (const ispc::vec3f &)c0,
+      (const ispc::vec3f &)intensityDistribution.c0,
       (const ispc::vec3f &)radiance,
       (const ispc::vec3f &)radIntensity,
       cosAngleMax,
       cosAngleScale,
       radius,
       innerRadius,
-      (const ispc::vec2i &)size,
-      lid ? lid->data() : nullptr);
+      (const ispc::vec2i &)intensityDistribution.size,
+      intensityDistribution.data());
   return ie;
 }
 
@@ -43,30 +43,10 @@ void SpotLight::commit()
       clamp(getParam<float>("innerRadius", 0.f), 0.0f, 0.999f * radius);
 
   // per default perpendicular to direction
-  c0 = std::abs(direction.x) < std::abs(direction.y)
+  intensityDistribution.c0 = std::abs(direction.x) < std::abs(direction.y)
       ? vec3f(0.0f, direction.z, direction.y)
       : vec3f(direction.z, 0.0f, direction.x);
-  c0 = getParam<vec3f>("c0", c0);
-  lid = getParamDataT<float, 2>("intensityDistribution");
-  if (lid) {
-    if (lid->numItems.z > 1)
-      throw std::runtime_error(toString()
-          + " must have (at most 2D) 'intensityDistribution' array using the first two dimensions.");
-    size = vec2i(lid->numItems.x, lid->numItems.y);
-    if (size.x < 2)
-      throw std::runtime_error(toString()
-          + " 'intensityDistribution' must have data for at least two gamma angles.");
-    if (!lid->compact()) {
-      postStatusMsg(OSP_LOG_WARNING)
-          << toString()
-          << " does currently not support strides for 'intensityDistribution', copying data.";
-
-      const auto data = new Data(OSP_FLOAT, lid->numItems);
-      data->copy(*lid, vec3ui(0));
-      lid = &(data->as<float, 2>());
-      data->refDec();
-    }
-  }
+  intensityDistribution.readParams(*this);
 
   // check ranges and pre-compute parameters
   openingAngle = clamp(openingAngle, 0.f, 360.f);
@@ -77,7 +57,9 @@ void SpotLight::commit()
       std::cos(deg2rad(0.5f * openingAngle - penumbraAngle));
   cosAngleScale = 1.0f / (cosAngleMin - cosAngleMax);
 
-  queryIntensityQuantityType(OSP_INTENSITY_QUANTITY_INTENSITY);
+  queryIntensityQuantityType(intensityDistribution
+          ? OSP_INTENSITY_QUANTITY_SCALE
+          : OSP_INTENSITY_QUANTITY_INTENSITY);
   processIntensityQuantityType(openingAngle);
 }
 
@@ -85,45 +67,42 @@ void SpotLight::processIntensityQuantityType(const float openingAngle)
 {
   radIntensity = 0.0f;
   radiance = 0.0f;
+  const float halfOpeningAngleRad = M_PI * (openingAngle * 0.5f) / 180.0f;
+  const float cosHalfOpeningAngle = cos(halfOpeningAngleRad);
+  const auto sqr = [](const float f) { return f * f; };
+  const float ringDiskArea = M_PI * (sqr(radius) - sqr(innerRadius));
+  const float sphericalCapCosInt = M_PI * (1.0f - sqr(cosHalfOpeningAngle));
 
-  float halfOpeningAngleRad = M_PI * (openingAngle * 0.5f) / 180.0f;
-  float cosHalfOpeningAngle = cos(halfOpeningAngleRad);
-  float ringDiskArea = M_PI * (radius * radius - innerRadius * innerRadius);
-
-  float sphericalCapCosInt =
-      M_PI * (1.0f - cosHalfOpeningAngle * cosHalfOpeningAngle);
-
-  if (intensityQuantity == OSP_INTENSITY_QUANTITY_INTENSITY) {
+  // converting from the chosen intensity quantity type to radiance
+  if (intensityDistribution
+          ? intensityQuantity == OSP_INTENSITY_QUANTITY_SCALE
+          : intensityQuantity == OSP_INTENSITY_QUANTITY_INTENSITY) {
     radIntensity = coloredIntensity;
-    if (radius > 0.0f) {
+    if (radius > 0.0f)
       radiance = radIntensity / ringDiskArea;
-    }
-  } else if (intensityQuantity == OSP_INTENSITY_QUANTITY_POWER) {
-    // since our spot light implementation includes the cosine term
-    // we need to consider the integrated cosine cap instead of the
-    // usually used integrated cap.
-    radIntensity = coloredIntensity / sphericalCapCosInt;
-    if (radius > 0.0f) {
-      radiance = coloredIntensity / (sphericalCapCosInt * ringDiskArea);
-    }
-    if (lid) {
-      static WarnOnce warning(
-          "The 'intensityQuantity' : 'OSP_INTENSITY_QUANTITY_POWER' is not supported when using an 'intensityDistribution'");
-      radIntensity = 0.0f;
-      radiance = 0.0f;
-    }
-  } else if (intensityQuantity == OSP_INTENSITY_QUANTITY_RADIANCE) {
-    // a virtual spot light has no surface area
-    // therefore radIntensity stays zero and radiance is only
-    // set if radius > 0
-    if (radius > 0.0f) {
-      radiance = coloredIntensity;
-    }
-  } else {
-    static WarnOnce warning(
-        "Unsupported intensityQuantity type for a 'spot' light source");
-    radiance = vec3f(0.0f);
+    return;
   }
+  if (!intensityDistribution) {
+    if (intensityQuantity == OSP_INTENSITY_QUANTITY_POWER) {
+      // since our spot light implementation includes the cosine term we need
+      // to consider the integrated cosine cap instead of the usually used
+      // integrated cap
+      radIntensity = coloredIntensity / sphericalCapCosInt;
+      if (radius > 0.0f)
+        radiance = coloredIntensity / (sphericalCapCosInt * ringDiskArea);
+      return;
+    }
+  }
+  // XXX support for RADIANCE with intensityDistribution is deprecated
+  if (intensityQuantity == OSP_INTENSITY_QUANTITY_RADIANCE) {
+    // a virtual spot light has no surface area therefore radIntensity stays 0
+    if (radius > 0.0f)
+      radiance = coloredIntensity;
+    return;
+  }
+
+  postStatusMsg(OSP_LOG_WARNING)
+      << toString() << " unsupported 'intensityQuantity' value";
 }
 
 } // namespace ospray