+++ modules/cpu/lights/SpotLight.h
@@ -1,10 +1,10 @@
-// Copyright 2009-2021 Intel Corporation
+// Copyright 2009-2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
 
+#include "IntensityDistribution.h"
 #include "Light.h"
-#include "common/Data.h"
 
 namespace ospray {
 
@@ -21,17 +21,15 @@ struct OSPRAY_SDK_INTERFACE SpotLight : public Light
  private:
   void processIntensityQuantityType(const float openingAngle);
 
-  Ref<const DataT<float, 2>> lid; // luminous intensity distribution
-  vec2i size{0};
   vec3f position{0.f};
   vec3f direction{0.f, 0.f, 1.f};
-  vec3f c0{1.f, 0.f, 0.f};
   vec3f radiance{1.f};
   vec3f radIntensity{0.f};
   float radius{0.f};
   float innerRadius{0.f};
   float cosAngleMax{1.f};
   float cosAngleScale{1.f};
+  IntensityDistribution intensityDistribution;
 };
 
 } // namespace ospray